api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
supabase==2.0.2
openai==1.3.0
python-dotenv==1.0.0
pydantic==2.5.0
python-docx==0.8.11
python-dateutil==2.8.2
requests==2.31.0
